containers/router-agent/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
langchain
langchain-google-vertexai
langchain-core
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
prometheus-fastapi-instrumentator
openinference-instrumentation-langchain
pyyaml
python-frontmatter
loguru
